Code refactoring

+++ src/replay.py
@@ -1,9 +1,8 @@
 import os, shutil, jinja2
 import pandas as pd
 import numpy as np
-import matplotlib.pyplot as plt
 from xml.dom import minidom
-from allcurves import *
+import allcurves as allcurves_code
 import logging, datetime
 
 # logging.basicConfig(
@@ -14,11 +13,18 @@ import logging, datetime
 
 
 def gen_table(csvfile, output_dir):
-    os.system("rm " + output_dir + "table.txt")
+    # Remove previous files
+    if os.path.isfile(output_dir + "table.txt"):
+        os.system("rm " + output_dir + "table.txt")
+
+    # Create new table file
     df = pd.read_csv(csvfile, sep=";")
     time = df.iloc[:, 0]
-    term_names = [x.split("_V_re")[0].replace(" ", "-") for x in df.columns[1:-1] if "V_re" in x]
-    omega_names = [x.replace(" ", "-") for x in df.columns[1:-1] if "omega" in x]
+    time = time - list(time)[0]
+    # term_names = [x.split("_V_re")[0].replace(" ", "-") for x in df.columns[1:-1] if "V_re" in x]
+    term_names = [x.split("_V_re")[0].replace(" ", " ") for x in df.columns[1:-1] if "V_re" in x]
+    omega_names = [x.replace(" ", " ") for x in df.columns[1:-1] if "omega" in x]
+    # omega_names = [x.replace(" ", "-") for x in df.columns[1:-1] if "omega" in x]
     with open(output_dir + "table.txt", "w") as f:
         f.write("#1\n")
         for omega in omega_names:
@@ -45,13 +51,19 @@ def get_jobs_config(jobsfile):
     simulation = file.getElementsByTagName("simulation")[0]
     system = {}
     system["simulation"] = {
-        "startTime": simulation.attributes["startTime"].value,
-        "stopTime": simulation.attributes["stopTime"].value,
+        "startTime": str(0),
+        "stopTime": str(
+            float(simulation.attributes["stopTime"].value)
+            - float(simulation.attributes["startTime"].value)
+        ),
+        "precision": simulation.attributes["precision"].value
+        if "precision" in simulation.attributes
+        else "1e-8",
     }
     return system
 
 
-def get_generators(dydfile, tagPrefix="dyn:"):
+def get_generators(dydfile, tagPrefix):
     file = minidom.parse(dydfile)
     models = file.getElementsByTagName(tagPrefix + "blackBoxModel")
     generators = [x for x in models if "Generator" in x.attributes["lib"].value]
@@ -91,13 +103,16 @@ def get_solver_params(jobsfile, parfile, system, tagPrefix=""):
     system["solver"] = [
         x for x in parsets if solver.attributes["parId"].value == x.attributes["id"].value
     ][0]
+    system["solver"].setAttribute("lib", solver.getAttribute("lib"))
 
 
 def get_refs(generators, parsets, iidmfile):
     if not os.path.isfile(iidmfile):
         return 0
     iidm = minidom.parse(iidmfile)
-    iidm_generators = iidm.getElementsByTagName("iidm:generator")
+    iidm_generators = iidm.getElementsByTagName("generator")
+    if len(iidm_generators) == 0:
+        iidm_generators = iidm.getElementsByTagName("iidm:generator")
     for s in parsets:
         refs = s.getElementsByTagName("reference")
         dyd_gen_id = [
@@ -157,14 +172,14 @@ This file is part of Dynawo, an hybrid C++/Modelica open source time domain
 simulation tool for power systems. -->
 <jobs xmlns="http://www.rte-france.com/dynawo">
     <job name="{{system['name']}}">
-        <solver lib="dynawo_SolverIDA" parFile="{{system['name']}}.par" parId="{{system['solver'].attributes['id'].value}}"/>
+        <solver lib="{{system['solver'].attributes['lib'].value}}" parFile="{{system['name']}}.par" parId="{{system['solver'].attributes['id'].value}}"/>-->
         <modeler compileDir="outputs/compilation">
             <!-- <network iidmFile="{{system['name']}}.iidm" parFile="{{system['name']}}.par" parId="Network"/> -->
             <dynModels dydFile="{{system['name']}}.dyd"/>
             <precompiledModels useStandardModels="true"/>
             <modelicaModels useStandardModels="true"/>
         </modeler>
-        <simulation startTime="{{system['simulation']['startTime']}}" stopTime="{{system['simulation']['stopTime']}}"/>
+        <simulation startTime="{{system['simulation']['startTime']}}" stopTime="{{system['simulation']['stopTime']}}" precision="{{system['simulation']['precision']}}"/>
         <outputs directory="outputs">
             <curves inputFile="{{system['name']}}.crv" exportMode="CSV"/>
 
@@ -177,9 +192,10 @@ simulation tool for power systems. -->
         f.write(template.render(system=system))
 
 
-def gen_dyd(system, output):
+def gen_dyd(system, output, tagPrefix):
     models = [x["dyd"] for x in system["generators"].values()]
-    template_src = """<?xml version='1.0' encoding='UTF-8'?>
+    template_src = (
+        """<?xml version='1.0' encoding='UTF-8'?>
 <!--
 Copyright (c) 2022, RTE (http://www.rte-france.com)
 See AUTHORS.txt
@@ -192,15 +208,28 @@ SPDX-License-Identifier: MPL-2.0
 This file is part of Dynawo, an hybrid C++/Modelica open source suite of
 simulation tools for power systems.
 -->
-<dyn:dynamicModelsArchitecture xmlns:dyn="http://www.rte-france.com/dynawo">
+<"""
+        + tagPrefix
+        + """dynamicModelsArchitecture xmlns:dyn="http://www.rte-france.com/dynawo">
     {% for model in models %}
-    <dyn:blackBoxModel id="{{model.attributes['id'].value}}" lib="{{model.attributes['lib'].value}}" parFile="{{system['name']}}.par" parId="{{model.attributes['parId'].value}}" />
-    <dyn:blackBoxModel id="IBus_{{model.attributes['id'].value}}" lib="InfiniteBusFromTable" parFile="{{system['name']}}.par" parId="IBus_{{model.attributes['id'].value}}"/>
-    <dyn:connect id1="{{model.attributes['id'].value}}" var1="generator_terminal" id2="IBus_{{model.attributes['id'].value}}" var2="infiniteBus_terminal"/>
-    <dyn:connect id1="{{model.attributes['id'].value}}" var1="generator_omegaRefPu_value" id2="IBus_{{model.attributes['id'].value}}" var2="infiniteBus_omegaRefPu"/>
+    <"""
+        + tagPrefix
+        + """blackBoxModel id="{{model.attributes['id'].value}}" lib="{{model.attributes['lib'].value}}" parFile="{{system['name']}}.par" parId="{{model.attributes['parId'].value}}"/>
+    <"""
+        + tagPrefix
+        + """blackBoxModel id="IBus_{{model.attributes['id'].value}}" lib="InfiniteBusFromTable" parFile="{{system['name']}}.par" parId="IBus_{{model.attributes['id'].value}}"/>
+    <"""
+        + tagPrefix
+        + """connect id1="{{model.attributes['id'].value}}" var1="generator_terminal" id2="IBus_{{model.attributes['id'].value}}" var2="infiniteBus_terminal"/>
+    <"""
+        + tagPrefix
+        + """connect id1="{{model.attributes['id'].value}}" var1="generator_omegaRefPu_value" id2="IBus_{{model.attributes['id'].value}}" var2="infiniteBus_omegaRefPu"/>
     {% endfor %}
-</dyn:dynamicModelsArchitecture>
+</"""
+        + tagPrefix
+        + """dynamicModelsArchitecture>
     """
+    )
     template = jinja2.Template(template_src)
     with open(output, "w") as f:
         f.write(template.render(system=system, models=models))
@@ -235,31 +264,136 @@ def gen_par(system, output):
     <par name="infiniteBus_UPuTableName"  type="STRING" value="{{modelpars.attributes['dydId'].value.replace(' ', '-')}}_generator_terminal_U"/>
     <par name="infiniteBus_UPhaseTableName" type="STRING" value="{{modelpars.attributes['dydId'].value.replace(' ', '-')}}_generator_terminal_UPhase"/>
     <par name="infiniteBus_OmegaRefPuTableName" type="STRING" value="{{modelpars.attributes['dydId'].value.replace(' ', '-')}}_generator_omegaRefPu_value"/>
-    <par name="infiniteBus_TableFile" type="STRING" value="table.txt"/>
+    <par name="infiniteBus_TableFile" type="STRING" value="{{infinite_bus_table}}"/>
   </set>
   {% endfor -%}
 
-   <set id="IDAOrder2">
-    <par type="INT" name="order" value="2"/>
-    <par type="DOUBLE" name="initStep" value="1e-9"/>
-    <par type="DOUBLE" name="minStep" value="1e-9"/>
-    <par type="DOUBLE" name="maxStep" value="1"/>
-    <par type="DOUBLE" name="absAccuracy" value="1e-6"/>
-    <par type="DOUBLE" name="relAccuracy" value="1e-6"/>
-    <par type="DOUBLE" name="minimalAcceptableStep" value="1e-10"/>
-    <par type="INT" name="maximumNumberSlowStepIncrease" value="40"/>
-  </set>
+   {{ system['solver'].toxml() }}
+   
 </parametersSet>
     """
     template = jinja2.Template(template_src)
     with open(output, "w") as f:
-        f.write(template.render(system=system, parlist=parlist))
+        f.write(
+            template.render(
+                system=system,
+                parlist=parlist,
+                infinite_bus_table=system["infinite_bus_table"],
+            )
+        )
 
 
+# TODO: Substitute the IDA solver for the first line above in order to get the original params from simulation
 # {{ system['solver'].toxml() }}
 # {% for param in modelpars.getElementsByTagName('reference') -%}
 # <par name="{{param.attributes['name'].value}}" type="{{param.attributes['type'].value}}" value="1"/>
 # {% endfor -%}
+
+
+def gen_par_IBus(system, output, parlist):
+
+    file = minidom.parse(output)
+    parametersSet = file.getElementsByTagName("parametersSet")[0]
+
+    set_list = file.getElementsByTagName("set")
+    for modelpars in parlist:
+        for set_elem in set_list:
+            if set_elem.getAttribute("id") == modelpars.attributes["id"].value:
+                set_elem.setAttribute("dydId", modelpars.attributes["dydId"].value)
+                break
+
+        IBus_par = file.createElement("set")
+        IBus_par.setAttribute("id", "IBus_" + modelpars.attributes["dydId"].value)
+
+        par1 = file.createElement("par")
+        par2 = file.createElement("par")
+        par3 = file.createElement("par")
+        par4 = file.createElement("par")
+
+        par1.setAttribute("name", "infiniteBus_UPuTableName")
+        par2.setAttribute("name", "infiniteBus_UPhaseTableName")
+        par3.setAttribute("name", "infiniteBus_OmegaRefPuTableName")
+        par4.setAttribute("name", "infiniteBus_TableFile")
+
+        par1.setAttribute("type", "STRING")
+        par2.setAttribute("type", "STRING")
+        par3.setAttribute("type", "STRING")
+        par4.setAttribute("type", "STRING")
+
+        par1.setAttribute(
+            "value",
+            modelpars.attributes["dydId"].value.replace(" ", "-") + "_generator_terminal_U",
+        )
+        par2.setAttribute(
+            "value",
+            modelpars.attributes["dydId"].value.replace(" ", "-") + "_generator_terminal_UPhase",
+        )
+        par3.setAttribute(
+            "value",
+            modelpars.attributes["dydId"].value.replace(" ", "-") + "_generator_omegaRefPu_value",
+        )
+        par4.setAttribute("value", system["infinite_bus_table"])
+
+        IBus_par.appendChild(par1)
+        IBus_par.appendChild(par2)
+        IBus_par.appendChild(par3)
+        IBus_par.appendChild(par4)
+
+        parametersSet.appendChild(IBus_par)
+
+    # Write output without whitespaces
+    with open(output, "w") as out:
+        xml_str = file.toprettyxml()
+        xml_str = os.linesep.join(
+            [
+                s
+                for s in xml_str.splitlines()
+                if s != "	" and s and s != "		" and s != "    " and s != "  "
+            ]
+        )
+        out.write(xml_str)
+        out.close()
+
+
+def solve_references(par_file, dumpinit_folder):
+    file = minidom.parse(par_file)
+    reference_list = file.getElementsByTagName("reference")
+
+    for reference_elem in reference_list:
+        if reference_elem.parentNode.hasAttribute("dydId"):
+            atr_ref = reference_elem.getAttribute("name")
+            value_ref = -9999999999
+            with open(
+                dumpinit_folder
+                + "dumpInitValues-{}.txt".format(reference_elem.parentNode.getAttribute("dydId"))
+            ) as f:
+                for line in f:
+                    line = line.rstrip()
+                    if line[: len(atr_ref)] == atr_ref and line[len(atr_ref)] == " ":
+                        value_ref = line.split("=")[1].replace(" ", "")
+                        break
+            if value_ref == -9999999999:
+                print(reference_elem.parentNode.getAttribute("dydId"), atr_ref)
+            par_ref = file.createElement("par")
+            par_ref.setAttribute("type", reference_elem.getAttribute("type"))
+            par_ref.setAttribute("name", atr_ref)
+            par_ref.setAttribute("value", str(value_ref))
+
+            reference_elem.parentNode.appendChild(par_ref)
+            reference_elem.parentNode.removeChild(reference_elem)
+
+    # Write output without whitespaces
+    with open(par_file, "w") as out:
+        xml_str = file.toprettyxml()
+        xml_str = os.linesep.join(
+            [
+                s
+                for s in xml_str.splitlines()
+                if s != "	" and s and s != "		" and s != "    " and s != "  "
+            ]
+        )
+        out.write(xml_str)
+        out.close()
 
 
 def gen_crv(system, output):
@@ -297,30 +431,57 @@ def remove_namespaces(file, tag):
     os.system("sed -i 's/<\\/{}:/<\\//g' {}".format(tag, file))
 
 
-def gen_replay_files(root_dir, model, terminals_csv, output_dir="replay/"):
+def gen_replay_files(root_dir, model, terminals_csv, output_dir, tagPrefix):
     os.makedirs(output_dir, exist_ok=True)
+
+    # Generate the replay table file
     gen_table(terminals_csv, output_dir)
-    # gen_table(csvfile, output_dir)
+
     in_jobs, in_par, in_dyd, in_iidm, in_crv = [
         root_dir + model + x for x in [".jobs", ".par", ".dyd", ".iidm", ".crv"]
     ]
-    out_jobs, out_par, out_dyd, out_crv = [
-        output_dir + model + x for x in [".jobs", ".par", ".dyd", ".crv"]
+    out_jobs, out_par, out_dyd, out_iidm, out_crv = [
+        output_dir + model + x for x in [".jobs", ".par", ".dyd", ".iidm", ".crv"]
     ]
+
     # remove_namespaces(in_dyd, 'dyn')
     # remove_namespaces(in_iidm, 'iidm')
+
+    # Get simulation config params
     system = get_jobs_config(in_jobs)
-    gen_dict = get_generators(in_dyd, "dyn:")
+
+    # Get list of all dyd gens
+    gen_dict = get_generators(in_dyd, tagPrefix)
+
+    # TODO: At the moment the original parameters are used, in
+    # the future, to save memory, only the generator parameters
+    # with these functions could be selected
+
+    # Add gen initilization params
     get_gen_params(in_par, gen_dict)
-    # solver_params = gen_dict['solver']
-    # gen_dict.pop('solver')
-    parlist = [x["par"] for x in gen_dict.values()]
-    get_refs(gen_dict, parlist, root_dir + model + ".iidm")
+
+    # Get initialization params that are defined in the iidm
+    parlist = list(dict.fromkeys([x["par"] for x in gen_dict.values()]))
+
+    # get_refs(gen_dict, parlist, root_dir + model + ".iidm")
+
     system["generators"] = gen_dict
     system["name"] = model
+    system["infinite_bus_table"] = output_dir + "table.txt"
+
+    # Get the solver parameters to use them in the recreation of the curves
     get_solver_params(in_jobs, in_par, system)
+
+    # Generate the replay simulation files with all the data obtained above
     gen_jobs(system, out_jobs)
-    gen_dyd(system, out_dyd)
-    gen_par(system, out_par)
-    os.system("cp {}*_replay.crv {}/{}.crv".format(root_dir, output_dir, model))
+    gen_dyd(system, out_dyd, tagPrefix)
+    os.system("cp '{}/{}.crv' '{}/{}.crv'".format(root_dir, model, output_dir, model))
+    os.system("cp '{}/{}.par' '{}/{}.par'".format(root_dir, model, output_dir, model))
+    # Add IBus to pars
+    # gen_par(system, out_par)
+    gen_par_IBus(system, out_par, parlist)
+    solve_references(out_par, os.path.dirname(terminals_csv) + "/../initValues/globalInit/")
+
+    # TODO: Study what curves should be replayed and modify this part
+
     return system
